requirements.txt
SpeechRecognition
pyttsx3
keyboard
requests
vosk
pystray
Pillow
rich
pyaudio
